requirements.txt
Click==7.0
suds-jurko==0.6
Werkzeug==0.15.5
Flask==0.12.4
Flask-Spyne==0.2
itsdangerous==1.1.0
Jinja2==2.10.1
lxml==4.4.0
MarkupSafe==1.1.1
pycryptodome==3.9.4
pytz==2019.2
PyYAML==5.1.2
sec-wall==1.2
spyne==2.12.16

